+++ src/bar.rs
@@ -55,6 +55,18 @@ impl<'a> Bar<'a> {
         }
     }
 
+    pub(crate) const fn empty() -> Self {
+        Self {
+            empty: None,
+            complete: None,
+            in_between: None,
+            left_border: None,
+            right_border: None,
+            filled_style: None,
+            empty_style: None,
+        }
+    }
+
     pub fn render(&self, width: usize, completed: f64) -> String {
         let completed = (completed * width as f64) as usize;
         let remaining = width.saturating_sub(completed);
@@ -87,25 +99,25 @@ impl<'a> Bar<'a> {
         )
     }
 
-    pub fn with_in_between(mut self, chars: &'static str) -> Self {
+    pub const fn with_in_between(mut self, chars: &'static str) -> Self {
         self.in_between = Some(chars);
         self
     }
 
-    pub fn with_border(mut self, left: &'static str, right: &'static str) -> Self {
+    pub const fn with_border(mut self, left: &'static str, right: &'static str) -> Self {
         self.left_border = Some(left);
         self.right_border = Some(right);
         self
     }
 
     /// Style the filled portion of the bar.
-    pub fn with_filled_style(mut self, style: owo_colors::Style) -> Self {
+    pub const fn with_filled_style(mut self, style: owo_colors::Style) -> Self {
         self.filled_style = Some(style);
         self
     }
 
     /// Style the empty portion of the bar.
-    pub fn with_empty_style(mut self, style: owo_colors::Style) -> Self {
+    pub const fn with_empty_style(mut self, style: owo_colors::Style) -> Self {
         self.empty_style = Some(style);
         self
     }